add getLastTransactions stub - [CORE-849]
