--- dapps/digital-twin-detail/src/components/overview/DigitaltwinResponse.interface.ts
@@ -1,14 +1,0 @@
-export interface DigitalTwinResponse {
-    total: number;
-    results: Result[];
-  }
-
-export interface Result {
-    address: string;
-    containers: any[];
-    created: number;
-    description: string;
-    name: string;
-    owner: string;
-    updated: number;
-  }
