+++ dapps/digital-twin-detail/tsconfig.json
@@ -8,7 +8,7 @@
   "extends": "../../tsconfig.json",
   "include": [
     "./src/**/*",
-    "./src/**/*.json"
+    "./src/**/*.json", "../../core/digital-twin-lib/src/interfaces/TransactionsResponse.interface.ts", "../../core/digital-twin-lib/src/interfaces/DigitaltwinResponse.interface.ts"
   ],
   "references": [
     {
